+++ developing/scripts/install.sh
@@ -1,6 +1,3 @@
 #!/bin/sh
-gem install tty-prompt
-gem install json
-gem install dotenv
 cargo install nj-cli
 cargo install wasm-pack